requirements.txt
playwright
pytest
google-generativeai
pydantic
requests
beautifulsoup4
pytest-playwright